Footprint editor: fix broken mirroring of custom pad shapes. Fixes #8649 https://gitlab.com/kicad/code/kicad/issues/8649

+++ pcbnew/board_design_settings.cpp
@@ -377,7 +377,7 @@ public:
         wxString                   oldPath = aConfig->GetPath();
         NETCLASSES::const_iterator nc = m_Pt_param->begin();
 
-        for( int index = 0; index <= m_Pt_param->GetCount(); ++index )
+        for( size_t index = 0; index <= m_Pt_param->GetCount(); ++index )
         {
             wxString    path = "";
             NETCLASSPTR netclass;
